🚑 rename: api와 service로 분리리
