+++ src/pages/mypage/components/guestbook/GuestbookModal.tsx
@@ -4,7 +4,7 @@ import Button from '@/components/common/Button';
 import Heart from '@/components/common/Heart';
 import ModalInput from '@/components/common/ModalInput';
 import { GuestbookItem } from '@/types';
-import { useCreateGuestbook, useUpdateGuestbook } from '@/utils/query/guestbook/myGuestbooks';
+import { useCreateGuestbook, useUpdateGuestbook } from '@/pages/mypage/service/myGuestbooks';
 import Preparing from '@/components/common/Preparing';
