+++ src/pages/mypage/components/join-gathering/JoinGathering.tsx
@@ -1,11 +1,5 @@
-// MyGatheringTab.tsx
-import { sortGatheringsByDate } from '@/utils/sortGatherings';
-import ChallengeSection from '../gathering-section/ChallengeSection';
-import MainCard from '../gathering-section/MainCard';
-import CanceledGathering from '@/components/common/CanceledGathering';
-import Null from '@/components/common/Null';
-import { useState } from 'react';
-import Preparing from '@/components/common/Preparing';
+import GatheringList from '@/pages/mypage/components/gathering-section/GatheringList';
+import { GatheringItem, GatheringStateType, GatheringChallengeType } from '@/types';
 import {
   userGatherings,
   userGatheringStates,
@@ -17,74 +11,15 @@ interface JoinGatheringProps {
   onCancelParticipation: (gatheringId: number) => void;
 }
 
-export default function JoinGathering({
-  onCancelParticipation,
-}: JoinGatheringProps) {
-  const [openChallenges, setOpenChallenges] = useState<{
-    [key: number]: boolean;
-  }>({});
-
-  const handleToggleChallenge = (gatheringId: number) => {
-    setOpenChallenges((prev) => ({
-      ...prev,
-      [gatheringId]: !prev[gatheringId],
-    }));
-  };
-
-  const validGatherings = userGatherings.filter(gathering => {
-    const state = userGatheringStates[gathering?.gatheringId];
-    return gathering && state;
-  });
-
-  if (validGatherings.length === 0) {
-    return <Null message="아직 참여한 모임이 없습니다." />;
-  }
-
-  const sortedGatherings = sortGatheringsByDate(validGatherings);
-
+export default function JoinGathering({ onCancelParticipation }: JoinGatheringProps) {
   return (
-    <div className="space-y-6 pb-[50px]">
-      {sortedGatherings.map((gathering) => {
-        const state = userGatheringStates[gathering.gatheringId];
-        if (!state) return null;
-
-        const challenges = userGatheringChallenges[gathering.gatheringId];
-        const isOpen = openChallenges[gathering.gatheringId];
-
-        return (
-          <div
-            key={gathering.gatheringId}
-            className="relative rounded-lg overflow-hidden mb-[50px]"
-          >
-            <Preparing isVisible={true} message="api 준비 중인 서비스입니다..." />
-            <MainCard
-              gathering={gathering}
-              state={state}
-              onCancelParticipation={onCancelParticipation}
-            />
-
-            <ChallengeSection
-              challenges={challenges}
-              gathering={gathering}
-              isOpen={isOpen}
-              onToggle={() => handleToggleChallenge(gathering.gatheringId)}
-            />
-
-            <CanceledGathering
-              type="gathering"
-              gatheringStartDate={gathering.gatheringStartDate}
-              gatheringJoinedPeopleCount={state.gatheringJoinedPeopleCount}
-              isReservationCancellable={gathering.isReservationCancellable || false}
-              onOverlay={() => {
-                setOpenChallenges((prev) => ({
-                  ...prev,
-                  [gathering.gatheringId]: false,
-                }));
-              }}
-            />
-          </div>
-        );
-      })}
-    </div>
+    <GatheringList
+      gatherings={userGatherings as GatheringItem[]}
+      gatheringStates={userGatheringStates as { [key: number]: GatheringStateType }}
+      gatheringChallenges={userGatheringChallenges as { [key: number]: GatheringChallengeType }}
+      emptyMessage="아직 참여한 모임이 없습니다."
+      onCancelAction={onCancelParticipation}
+      cancelActionType="participation"
+    />
   );
 }
